src/multi_node/requirements.txt
# Requirements for Multi-Node Temperature Monitoring System

dash>=2.14.0
plotly>=5.15.0
paho-mqtt>=1.6.0
numpy>=1.24.0

# Core Python libraries (usually included)
# sqlite3 - included in Python standard library
# threading - included in Python standard library
# json - included in Python standard library
# datetime - included in Python standard library
# collections - included in Python standard library
# sys - included in Python standard library
# random - included in Python standard library
# time - included in Python standard library
# math - included in Python standard library